issue #169 feat: Common Tab add
