+++ src/components/Header/SubHeader/SubHeader.jsx
@@ -1,26 +1,13 @@
 import React from "react";
-import { useNavigate, useLocation } from "react-router-dom";
 
-import { SubHeaderDiv, SubTabUl, TabButton } from "./SubHeader.style";
+import Tab from "@/components/Base/Tab/Tab";
 
-const SubHeader = ({ tabData }) => {
-	const navigate = useNavigate();
-	const location = useLocation();
+import { SubHeaderDiv } from "./SubHeader.style";
 
+const SubHeader = ({ data }) => {
 	return (
 		<SubHeaderDiv>
-			<SubTabUl role="tablist">
-				{tabData.map(({ id, title, link }) => (
-					<li key={id} role="tab">
-						<TabButton
-							onClick={() => navigate(link)}
-							isActive={location.pathname.includes(link)}
-						>
-							{title}
-						</TabButton>
-					</li>
-				))}
-			</SubTabUl>
+			<Tab data={data} name="subheader" />
 		</SubHeaderDiv>
 	);
 };
